moved code to handle new responses to appropriate files
